requirements.txt
# Core dependencies
python-dotenv==1.0.1
pinecone-client>=5.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-pinecone>=0.1.0
langchain-huggingface>=0.1.0
langchain-openai>=0.1.0

# Machine Learning and NLP
transformers==4.45.2
sentence-transformers==3.0.1
huggingface-hub==0.36.0
accelerate==0.34.2

# Document Processing
pypdf==4.3.1

# Frontend (optional)
gradio>=4.19.2  # Only needed if using the frontend

# Testing and Development
pytest>=7.0.0  # For running tests
black>=23.0.0  # For code formatting
isort>=5.0.0   # For import sorting